refactor: [YSS-82] dto 통합
