+++ yakssok/src/main/java/server/yakssok/domain/friend/presentation/dto/response/FollowingMedicationStatusResponse.java
@@ -1,5 +1,6 @@
 package server.yakssok.domain.friend.presentation.dto.response;
 
+import java.time.LocalTime;
 import java.util.List;
 
 import io.swagger.v3.oas.annotations.media.Schema;
@@ -26,6 +27,16 @@ public record FollowingMedicationStatusResponse(
 	@Schema(description = "잔소리 대상의 안먹은 약 / 칭찬 대상의 먹은 약 상세 정보")
 	List<MedicationInfo> medicationDetails
 ) {
+
+	public record MedicationInfo(
+		@Schema(description = "약 종류", example = "CHRONIC")
+		String type,
+		@Schema(description = "약 이름", example = "타이레놀")
+		String name,
+		@Schema(description = "복용 시간 (HH:mm:ss)", type = "string", format="HH:mm:ss", example = "08:00:00")
+		LocalTime time
+	) { }
+
 	/* ===== 공통 생성기 ===== */
 	private static FollowingMedicationStatusResponse ofCommon(
 		Friend friend,
